requirements.txt
dnspython
enum-compat
eventlet
greenlet
monotonic
python-engineio
python-socketio
pytz
six
mysqlclient
djangorestframework
pyJWT
djangorestframework-simplejwt
django-phonenumber-field[phonenumbers]
django-countries
django-cities
django-cors-headers
django-suit